+++ pyFAT_astro/Support/modify_template.py
@@ -1765,6 +1765,67 @@ get_number_of_rings.__doc__ =f'''
  NOTE:
 '''
 
+def get_ring_weights(Configuration,Tirific_Template):
+    print_log(f'''GET_RING_WEIGTHS: Getting the importance of the rings in terms of SBR.
+''',Configuration,case=['debug_start'])
+    sbr = load_tirific(Configuration,Tirific_Template,Variables=["SBR",f"SBR_2"],\
+                array=True )
+    
+    print_log(f'''GET_RING_WEIGTHS: retrieve this sbr.
+{'':8s} sbr = {sbr}
+''',Configuration,case=['debug_add'])
+    cut_off_limits = sf.sbr_limits(Configuration,Tirific_Template )
+    print_log(f'''GET_RING_WEIGTHS: retrieved these cut_off_limits.
+{'':8s} col = {cut_off_limits}
+''',Configuration,case=['debug_add'])
+    sm_sbr = smooth_profile(Configuration,Tirific_Template,'SBR',
+                            min_error= [cut_off_limits,cut_off_limits],no_apply = True,
+                            fix_sbr_call = True, profile_in = sbr )
+    print_log(f'''GET_RING_WEIGTHS: retrieve this sbr.
+{'':8s} sbr = {sbr}
+{'':8s} using sm_sbr {sm_sbr}
+''',Configuration,case=['debug_add'])
+
+    weights= [[],[]]
+    for i in [0,1]:
+        weights[i] = [sf.set_limits(x/y,0.1,10.) for x,y in zip(sm_sbr[i],cut_off_limits)]
+        weights[i] = weights[i]/np.nanmax(weights[i])
+        weights[i][0:2] = np.nanmin(weights[i])
+        weights[i] = [sf.set_limits(x,0.1,1.) for x in weights[i]]
+    print_log(f'''GET_RING_WEIGTHS: Obtained the following weights.
+{'':8s}{weights}
+''',Configuration,case=['debug_add'])
+    return np.array(weights,dtype = float)
+
+get_ring_weights.__doc__=f'''
+ NAME:
+    get_ring_weights
+
+ PURPOSE:
+    Get the importance of the rings based on how much the SBR lies above the noise limit for each ring
+
+ CATEGORY:
+    support_functions
+
+ INPUTS:
+    Configuration = Standard FAT configuration
+    Tirific_Template = Standard Tirific template containg the SBR profiles
+
+ OPTIONAL INPUTS:
+
+
+ OUTPUTS:
+    numpy array with the weight normalized to the the maximum.
+
+ OPTIONAL OUTPUTS:
+
+ PROCEDURES CALLED:
+    Unspecified
+
+ NOTE:
+    Weight 1 is most important, weight 0. least important.
+    Errors should be divided by these weights to reflect the importance
+'''
 def get_warp_slope(Configuration,Tirific_Template):
     print_log(f'''GET_WARP_SLOPE: We have {Tirific_Template['NUR']} rings in the template. and this should be {Configuration['NO_RINGS']}
 ''', Configuration,case=['debug_start'])
@@ -2087,7 +2148,7 @@ def regularise_profile(Configuration,Tirific_Template, key,min_error= None, \
         raise FunctionCallError('The warp is regularised in regularise_warp. ')
     # We start by getting an estimate for the errors
     min_error=np.array(min_error,dtype=float)
-    weights = sf.get_ring_weights(Configuration,Tirific_Template)
+    weights = get_ring_weights(Configuration,Tirific_Template)
     #For a lot of things it is important to use a unmodified input profile but we have to avoid the declining vrot
     if key == 'VROT':
         no_declining_vrot(Configuration, Tirific_Template)
@@ -2240,7 +2301,7 @@ def regularise_warp(Configuration,Tirific_Template, min_error = None, \
             min_error = [0.,0.]
 
     min_error=np.array(min_error,dtype=float)
-    weights = sf.get_ring_weights(Configuration,Tirific_Template)
+    weights = get_ring_weights(Configuration,Tirific_Template)
    
     #if the PA and INCL are fixed we stop here
     if 'PA' in Configuration['FIXED_PARAMETERS'][0] and 'INCL' in \
@@ -4167,7 +4228,7 @@ def smooth_profile(Configuration,Tirific_Template,key,min_error = 0.  \
 {'':8s}{profile}
 ''',Configuration,case= ['debug_add'])
     if not no_apply:
-        weights = sf.get_ring_weights(Configuration,Tirific_Template)
+        weights = get_ring_weights(Configuration,Tirific_Template)
         errors = get_error(Configuration,original_profile,profile,key,weights = weights, min_error=min_error)
         if key not in ['VROT']:
             # Check whether it should be flat
